Move the JobLogsDBConnectionTest to test_jobs
